backend/requirements.txt
# Este arquivo é atualizado automaticamente por update_requirements.py
# Evite editar manualmente. Instale bibliotecas no container e gere o freeze.

# === Core da API ===
fastapi==0.110.1            # Framework principal da API
uvicorn==0.29.0             # Servidor ASGI leve e rápido
starlette==0.37.2           # Base do FastAPI para rotas, middleware, etc.

# === Segurança / Autenticação ===
passlib==1.7.4              # Hashing seguro de senhas (bcrypt)
bcrypt==4.0.1               # Requisito interno do passlib
python-jose==3.3.0          # JWT (JSON Web Tokens) para autenticação

# === MongoDB ===
pymongo==4.7.1              # Driver oficial do MongoDB
dnspython==2.7.0            # Necessário para URIs SRV (Mongo Atlas)

# === Validações ===
email_validator==2.2.0      # Validação real de e-mail
pydantic==2.6.4             # Validação de entrada com tipagem forte
pydantic_core==2.16.3       # Núcleo do Pydantic
annotated-types==0.7.0      # Tipagem estendida para validações
typing-inspection==0.4.1    # Inspeção de tipos (útil para validações avançadas)
typing_extensions==4.13.2   # Compatibilidade com tipagem nova

# === Uploads / Formulários ===
python-multipart==0.0.9     # Necessário para processar formulários (ex: login)

# === HTTP (opcional para testes e integração) ===
httpx==0.27.0
httpcore==1.0.9
h11==0.16.0
sniffio==1.3.1
anyio==4.9.0
idna==3.10

# === Testes automatizados ===
pytest==8.2.1               # Test runner padrão do Python moderno
iniconfig==2.1.0            # Parser de config para pytest
pluggy==1.6.0               # Plugin manager interno usado pelo pytest
colorama==0.4.6             # Suporte a cores no terminal para logs

# === Outras dependências internas ===
certifi==2025.4.26          # Certificados SSL atualizados
click==8.2.1                # CLI usada por várias libs (opcional)
packaging==25.0             # Para manipular versões e empacotamento
six==1.17.0                 # Compatibilidade Python 2/3 (algumas libs ainda dependem)
